Refactor online test fixtures into separate conftest module
